Consolidate files and simplify
