+++ platformio.ini
@@ -893,6 +893,24 @@ build_flags =
   '-DTimeLedON=0.1'
   '-DLED_SEND_RECEIVE_ON=1'
 
+[env:heltec-display]   ; Heltec ESP32 Board with SSD1306 display
+platform = ${com.esp32_platform}
+board = heltec_wifi_lora_32_V2
+lib_deps =
+  ${com-esp.lib_deps}
+  ${libraries.wifimanager32}
+  ${libraries.heltec} 
+build_flags =
+  ${com-esp.build_flags}
+  ${test_env.build_flags}
+; *** OpenMQTTGateway Config ***
+  '-DGateway_Name="OpenMQTTGateway_Heltec_Display"'
+; *** OpenMQTTGateway Modules ***
+  '-DZboardHELTEC="HELTEC"'
+; *** Heltec Board Options ***
+  '-DLOG_TO_LCD=true'         ; Enable log to LCD
+;  '-DLOG_LEVEL_LCD=LOG_LEVEL_TRACE'
+
 [env:esp32dev-mqtt-fw-test]
 platform = ${com.esp32_platform}
 board = esp32dev
